# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Gapak API
  version: 1.0.0
  description: |
    Gapak is a privacy-first social network built around trust, layered identity,
    controlled visibility, secure media, live experiences, and time-based social interaction.

    Privacy is a core mechanic rather than a decorative setting. The API is designed for
    trust-aware access rules, metadata minimization, secure media delivery, and live social primitives.
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    accessCookie:
      type: apiKey
      in: cookie
      name: gapak_at
  schemas:
...
    UploadIntentRequest:
      type: object
      required: [fileName, mimeType, sizeBytes]
      properties:
        purpose: { type: string, enum: [POST_ATTACHMENT, STORY, PROFILE, TRUST_ROOM, LIVE_REPLAY] }
        fileName: { type: string }
        mimeType: { type: string }
        sizeBytes: { type: integer }
        checksumSha256: { type: string }
        multipart: { type: boolean }
        partSizeBytes: { type: integer }
    CompleteUploadRequest:
      type: object
      required: [parts]
      properties:
        parts:
          type: array
          items:
            type: object
            required: [partNumber, etag, sizeBytes]
            properties:
              partNumber: { type: integer }
              etag: { type: string }
              sizeBytes: { type: integer }
...
paths:
...
  /api/v1/media/upload-intents:
    post:
      tags: [Media / Video / File]
      security: [{ bearerAuth: [] }]
      summary: Legacy upload intent entrypoint for secure media flow
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UploadIntentRequest' }
      responses:
        '201': { description: Upload intent created }
  /api/v1/media/upload-sessions:
    post:
      tags: [Media / Video / File]
      security: [{ bearerAuth: [] }]
      summary: Initialize secure resumable upload session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UploadIntentRequest' }
      responses:
        '201': { description: Upload session initialized }
  /api/v1/media/upload-sessions/{sessionId}:
    get:
      tags: [Media / Video / File]
      security: [{ bearerAuth: [] }]
      summary: Get multipart upload session state
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Upload session }
  /api/v1/media/upload-sessions/{sessionId}/parts:
    post:
      tags: [Media / Video / File]
      security: [{ bearerAuth: [] }]
      summary: Request signed part grant for direct-to-storage upload
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Signed part request }
  /api/v1/media/upload-sessions/{sessionId}/complete:
    post:
      tags: [Media / Video / File]
      security: [{ bearerAuth: [] }]
      summary: Complete upload and enqueue video/media processing pipeline
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteUploadRequest' }
      responses:
        '200': { description: Upload completed and queued }
  /api/v1/media/upload-sessions/{sessionId}/abort:
    post:
      tags: [Media / Video / File]
      security: [{ bearerAuth: [] }]
      summary: Abort secure upload session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Upload session aborted }
  /api/v1/media/assets/{mediaId}:
    get:
      tags: [Media / Video / File]
      security: [{ bearerAuth: [] }]
      summary: Get secure media asset details with processed video and thumbnail metadata
      parameters:
        - in: path
          name: mediaId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Media asset details }
  /api/v1/media/assets/{mediaId}/playback-grants:
    post:
      tags: [Media / Video / File]
      security: [{ bearerAuth: [] }]
      summary: Issue short-lived playback access grant for protected media
      parameters:
        - in: path
          name: mediaId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201': { description: Playback grant issued }
  /api/v1/media/{mediaId}/finalize:
    post:
      tags: [Media / Video / File]
      security: [{ bearerAuth: [] }]
      summary: Legacy finalize endpoint for uploaded asset
      parameters:
        - in: path
          name: mediaId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Upload finalized }
  /api/v1/media/{mediaId}/access:
    get:
      tags: [Media / Video / File]
      security: [{ bearerAuth: [] }]
      summary: Legacy signed access data endpoint
      parameters:
        - in: path
          name: mediaId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Access data }
...